Fixing a bug with multiple :ref: not being resolved
Before this commit, if there were 2 :ref: to the same spot, because
"references" were not stored uniquely on Environment, we only resolved
the first, not the second. But, if we simply stored $this->dependencies
as a unique array on Environment, we could have a collision between
a :doc: and a :ref: with the same name. The hack of prefixing the
unresolved dependencies was added to work around this.