Persist meta's as cache after rendering
In my previous commit I believed that the meta's could be cached after
parsing had completed but the unit tests quickly proved me wrong.

During the rendering phase will the dependencies be resolved and
persisted in the meta's collection. This means that when we try to
persist the cache too soon, that this information is lost.

It might be worth investigating if there is a way to resolve this
dependencies before rendering; but that is beyond scope of this change.