Update events.rst
A few stylistic changes that IMHO might improve the document.
One particular change, is the use of the verb "invoke" and its past participle "invoked" when it comes to running the event handler. 
This is to follow collocations: dispatch an event, invoke an event handler, register/add an event listener, etc.