[TASK] Compare lines in HTML functional tests trimmed
Adjusting the whitespace in HTML files to what our renderer outputs is tiresome.

And nice formating of the html is really not in the scope of this project.  Multiple whitespaces are ignored in HTML anyway. So I would suggest to trim the lines before comparing instead of rtrimming them.

Adjusted 2 Tests to demonstrate this is wokring.

A follow up which compares the HTML structure in a more logical way would be nice but having this already makes testing much easier without giving up acuracy