[BUGIX] Do not register anonymous links
Anonymous links must not be registered as they have not name. Right now they are registered with their title.

This leads to multiple registrations of the same name eventually.

[BUGIX] Do not register anonymous links

Anonymous links must not be registered as they have not name. Right now they are registered with their title.

This leads to multiple registrations of the same name eventually.