[FEATURE] Allow multiple whitespaces as separator in directives and links
In the TYPO3 docs we like to use 2 white spaces after the two dots `..  `
marking the start of a directive or link. This is achived in the editors by
pressing tab after the two dots. This way directive and their blocks are
aligned when using 4 spaces for indentation:

```
..  my-directive:: something
    My directives content
```

Currently such a directive or link is not detected to be a directive.
However RST Syntax allows this: https://docutils.sourceforge.io/0.4/docs/ref/rst/restructuredtext.html#whitespace

For the time being tabs are not possible as blocks indented with tabs are not recognized.
Tabs are allowed but not recommended by the rst syntax. Shall we disallow tabs and maybe
issue a warning when a file contains tabs?