[TASK] Delegate escaping to Twig
After this discussion in the Symfony documentation channel: https://symfony-devs.slack.com/archives/C3ASP9LJJ/p1673725412479199

Added tests that XSS is not possible. Escaping of output should be done as late as possible, usually in the twig template. This way it is always done in the same place and not forgotten. When no twig is used we have to do the escaping right before the HTML is combined.